requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
PyYAML>=6.0
anthropic>=0.39.0
python-dotenv>=1.0.0
